+++ packages/moonstone/Scrollable/useKey.js
@@ -1,0 +1,17 @@
+import {useMonitorEvents} from './useMonitorEvents';
+
+const useKey = () => {
+	const {
+		deleteMonitorEventTarget,
+		setMonitorEventTarget
+	} = useMonitorEvents();
+
+	return {
+		deleteMonitorEventTarget,
+		setMonitorEventTarget
+	};
+}
+
+export {
+	useKey
+};
