+++ packages/moonstone/VirtualList/useRestoreSpotlight.js
@@ -1,11 +1,19 @@
 import Spotlight from '@enact/spotlight';
 import {useEffect, useRef} from 'react';
 
-const useRestoreSpotlight = () => {
+const getNumberValue = (index) => index | 0;
+
+const useRestoreSpotlight = (instance) => {
+    const {
+        uiRefCurrent
+    } = instance.currnet;
+
 	const variables = useRef({
 		preservedIndex: false,
 		restoreLastFocused: false
-	});
+    });
+
+    const containerNode = uiRefCurrent && uiRefCurrent.containerRef && uiRefCurrent.containerRef.current || null;
 
 	// componentDidUpdate
 	useEffect(restoreFocus);	// TODO : Handle exhaustive-deps ESLint rule.
